Refactor: シェアボタンのテストをビューテストから新たなシステムテストに切り分けた
