+++ spec/views/sakes/show.html.erb_spec.rb
@@ -94,61 +94,4 @@ RSpec.describe "sakes/show", type: :system do
       end
     end
   end
-
-  describe "share button" do
-    let(:sake) { create(:sake) }
-
-    before do
-      visit sake_path(sake.id)
-    end
-
-    it "has hidden accessibility text" do
-      within(:test_id, "share_button") do
-        hidden_elem = find(".visually-hidden")
-        expect(hidden_elem).to have_text(I18n.t("sakes.show_abstract.share_button"))
-      end
-    end
-
-    context "without sake informations" do
-      let(:no_info_sake) { create(:sake, name: "生道井") }
-
-      before do
-        visit sake_path(no_info_sake.id)
-      end
-
-      it "has hidden share text" do
-        target = find(:test_id, "share_text", visible: false).value
-        expected = "生道井。 #SAKAZUKI"
-        expect(target).to eq(expected)
-      end
-    end
-
-    context "with informations" do
-      let(:info_sake) {
-        create(:sake,
-               name: "生道井",
-               kura: "原田酒造合資会社",
-               todofuken: "愛知県",
-               color: "わずかに黄色",
-               aroma_impression: "りんご系のいい香り！",
-               taste_impression: "うまい。")
-      }
-
-      before do
-        visit sake_path(info_sake.id)
-      end
-
-      it "has hidden share text" do
-        target = find(:test_id, "share_text", visible: false).value
-        # 「。」「さんの」が追加される。「！」のあとは「。」が追加されない。
-        expected = "原田酒造㈾#{I18n.t("helper.share.honorific")}" +
-                   "生道井。" +
-                   "わずかに黄色#{I18n.t("helper.share.period")}" +
-                   "りんご系のいい香り！" +
-                   "うまい。" +
-                   " #SAKAZUKI"
-        expect(target).to have_text(expected)
-      end
-    end
-  end
 end
